Add add_sfml_demo cmake function in demo CMakeLists
Linked: #59
