+++ src/ecstasy/integrations/sfml/demo/02_events/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(INCROOT ${INCROOT}/02_events)
+set(SRCROOT ${SRCROOT}/02_events)
+
+
+set(SRC
+    ${SRCROOT}/main.cpp)
+
+add_sfml_demo(
+    NAME events
+    SOURCES ${SRC}
+    INCLUDE_DIRS ${INCROOT})
